@@ -1,53 +1,66 @@
 {
   "name": "frontend",
-  "version": "0.0.0",
+  "version": "2.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --host 0.0.0.0 --poll 500",
-    "start:proxy": "ng serve --host 0.0.0.0 --poll 500 --proxy-config proxy.conf.json",
-    "build": "ng build",
-    "watch": "ng build --watch --configuration development",
+    "start": "ng serve --proxy-config proxy.conf.json --host 0.0.0.0 --watch --poll 2000",
+    "serve": "ng serve --proxy-config proxy.conf.json",
+    "build": "ng build --configuration production",
     "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
     "dev:ssr": "ng run frontend:serve-ssr",
     "serve:ssr": "node dist/frontend/server/main.js",
-    "build:ssr": "ng build && ng run frontend:server",
-    "prerender": "ng run frontend:prerender"
+    "build:prod": "ng build --configuration production --output-path=dist",
+    "build:ssr": "ng build --configuration production && ng run frontend:server:production",
+    "build:output": "npm run build:ssr --output-path=dist",
+    "prerender": "ng run frontend:prerender",
+    "dev-server": "concurrently \"nodemon ../api/server.js --watch api\" \"npm run serve\""
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.6",
-    "@angular/common": "~12.2.6",
-    "@angular/compiler": "~12.2.6",
-    "@angular/core": "~12.2.6",
-    "@angular/forms": "~12.2.6",
-    "@angular/platform-browser": "~12.2.6",
-    "@angular/platform-browser-dynamic": "~12.2.6",
-    "@angular/platform-server": "~12.2.6",
-    "@angular/router": "~12.2.6",
-    "@nguniversal/express-engine": "^12.1.0",
+    "@angular/animations": "^14.2.2",
+    "@angular/common": "^14.2.2",
+    "@angular/compiler": "^14.2.2",
+    "@angular/core": "^14.2.2",
+    "@angular/forms": "^14.2.2",
+    "@angular/platform-browser": "^14.2.2",
+    "@angular/platform-browser-dynamic": "^14.2.2",
+    "@angular/platform-server": "^14.2.2",
+    "@angular/router": "^14.2.2",
+    "@nguniversal/express-engine": "^14.2.0",
     "@swimlane/ngx-datatable": "^20.0.0",
-    "bootstrap": "^5.1.1",
+    "bootstrap": "^4.5.3",
     "express": "^4.15.2",
-    "ngx-bootstrap": "^7.1.0",
-    "ngx-toastr": "^14.1.3",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "font-awesome": "^4.7.0",
+    "ngx-bootstrap": "^9.0.0",
+    "ngx-toastr": "^15.0.0",
+    "rxjs": "^6.6.7",
+    "tslib": "^2.3.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.2.6",
-    "@angular/cli": "~12.2.5",
-    "@angular/compiler-cli": "~12.2.6",
-    "@nguniversal/builders": "^11.0.1",
-    "@types/express": "^4.17.0",
-    "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.11.1",
-    "jasmine-core": "~3.8.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.3.5"
+    "@angular-devkit/build-angular": "^14.2.3",
+    "@angular/cli": "^14.2.3",
+    "@angular/compiler-cli": "^14.2.2",
+    "@angular/language-service": "^14.2.2",
+    "@nguniversal/builders": "^14.2.0",
+    "@types/express": "^4.17.14",
+    "@types/jasmine": "^4.3.0",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^18.7.18",
+    "codelyzer": "^6.0.2",
+    "concurrently": "^7.4.0",
+    "jasmine-core": "^4.4.0",
+    "jasmine-spec-reporter": "^7.0.0",
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "protractor": "~7.0.0",
+    "ts-node": "^10.9.1",
+    "tslint": "^6.1.3",
+    "typescript": "^4.8.3",
+    "zone.js": "^0.11.8"
   }
 }
