@@ -1,36 +1,42 @@
-import 'zone.js/dist/zone-node';
+import "zone.js/node";
 
-import { ngExpressEngine } from '@nguniversal/express-engine';
-import * as express from 'express';
-import { join } from 'path';
+import { ngExpressEngine } from "@nguniversal/express-engine";
+import * as express from "express";
+import { join } from "path";
 
-import { AppServerModule } from './src/main.server';
-import { APP_BASE_HREF } from '@angular/common';
-import { existsSync } from 'fs';
+import { AppServerModule } from "./src/main.server";
+import { APP_BASE_HREF } from "@angular/common";
+import { existsSync } from "fs";
 
 // The Express app is exported so that it can be used by serverless Functions.
 export function app(): express.Express {
   const server = express();
-  const distFolder = join(process.cwd(), 'dist/frontend/browser');
-  const indexHtml = existsSync(join(distFolder, 'index.original.html')) ? 'index.original.html' : 'index';
+  const distFolder = join(process.cwd(), "dist/frontend/browser");
+  const indexHtml = existsSync(join(distFolder, "index.original.html")) ? "index.original.html" : "index";
 
   // Our Universal express-engine (found @ https://github.com/angular/universal/tree/master/modules/express-engine)
-  server.engine('html', ngExpressEngine({
-    bootstrap: AppServerModule,
-  }));
+  server.engine(
+    "html",
+    ngExpressEngine({
+      bootstrap: AppServerModule
+    })
+  );
 
-  server.set('view engine', 'html');
-  server.set('views', distFolder);
+  server.set("view engine", "html");
+  server.set("views", distFolder);
 
   // Example Express Rest API endpoints
   // server.get('/api/**', (req, res) => { });
   // Serve static files from /browser
-  server.get('*.*', express.static(distFolder, {
-    maxAge: '1y'
-  }));
+  server.get(
+    "*.*",
+    express.static(distFolder, {
+      maxAge: "1y"
+    })
+  );
 
   // All regular routes use the Universal engine
-  server.get('*', (req, res) => {
+  server.get("*", (req, res) => {
     res.render(indexHtml, { req, providers: [{ provide: APP_BASE_HREF, useValue: req.baseUrl }] });
   });
 
@@ -52,9 +58,9 @@ function run(): void {
 // The below code is to ensure that the server is run only when not requiring the bundle.
 declare const __non_webpack_require__: NodeRequire;
 const mainModule = __non_webpack_require__.main;
-const moduleFilename = mainModule && mainModule.filename || '';
-if (moduleFilename === __filename || moduleFilename.includes('iisnode')) {
+const moduleFilename = (mainModule && mainModule.filename) || "";
+if (moduleFilename === __filename || moduleFilename.includes("iisnode")) {
   run();
 }
 
-export * from './src/main.server';
+export * from "./src/main.server";
